some fixed with test

+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: gpecci <gpecci@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/20 15:52:48 by gpecci            #+#    #+#             */
-/*   Updated: 2023/01/23 13:16:07 by gpecci           ###   ########.fr       */
+/*   Updated: 2023/01/23 14:51:19 by gpecci           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (s[i])
